+++ NEWS.md
@@ -2,6 +2,7 @@
 * automatically rename variables with conflicting capitalization (e.g., emax and EMAX)
 * rename `pk_elimination_linear_mm()` to `pk_elimination_linear_nl()`
 * remove `pk_elimination_mm()`
+* extended help for PK component building blocks
 
 # assemblerr 0.0.0.9008
 * extended help for the building blocks: parameter, observation, compartment, flow